requirements.txt
asgiref==3.3.4
Django==3.2
django-mathfilters==1.0.0
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
pytest==6.2.4
pytest-django==4.2.0
factory-boy==3.2.0

pip~=21.1.1
py~=1.10.0
setuptools~=56.2.0
Faker~=8.2.0